move validation of simulation time and draw functions to validation.py
